backend/pyproject.toml
[project]
name = "memories-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.2.1",
    "cloudinary>=1.44.1",
    "fastapi>=0.120.0",
    "gdown>=5.2.0",
    "google-genai>=1.46.0",
    "opencv-python>=4.12.0.88",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "tensorboard>=2.20.0",
    "torchreid>=0.2.5",
    "ultralytics>=8.3.217",
]
